+++ devview-networkmock/src/commonMain/kotlin/com/worldline/devview/networkmock/utils/ModelUtils.kt
@@ -10,6 +10,7 @@ import androidx.compose.material.icons.rounded.CloudOff
 import androidx.compose.material.icons.rounded.ErrorOutline
 import androidx.compose.material.icons.rounded.Info
 import androidx.compose.material.icons.rounded.Wifi
+import androidx.compose.material.icons.rounded.WifiOff
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.ReadOnlyComposable
@@ -77,6 +78,7 @@ internal fun OperationUiModel.Companion.fake(
 internal val OperationMockState.icon: ImageVector
     get() = when (this) {
         is OperationMockState.Mock -> iconForStatusCode(statusCode = statusCode)
+        is OperationMockState.Failure -> Icons.Rounded.WifiOff
         OperationMockState.Network -> Icons.Rounded.Wifi
     }
 
@@ -94,6 +96,7 @@ internal val OperationMockState.contentColor: Color
     @ReadOnlyComposable
     get() = when (this) {
         is OperationMockState.Mock -> contentColorForStatusCode(statusCode = statusCode)
+        is OperationMockState.Failure -> rememberMockColorScheme().failure.content
         OperationMockState.Network -> rememberMockColorScheme().network.content
     }
 
@@ -107,6 +110,7 @@ internal val OperationMockState.containerColor: Color
     @ReadOnlyComposable
     get() = when (this) {
         is OperationMockState.Mock -> containerColorForStatusCode(statusCode = statusCode)
+        is OperationMockState.Failure -> rememberMockColorScheme().failure.container
         OperationMockState.Network -> rememberMockColorScheme().network.container
     }